Aplicando patrones arquitectónicos a mi proyecto

+++ Exp1_S3_Florencia_Morice/TiendaRopaOnline/src/main/java/com/tienda/discount/Component.java
@@ -1,5 +1,7 @@
 package com.tienda.discount;
 
+import java.math.BigDecimal;
+
 public interface Component {
-    double aplicarDescuento(double precioBase);
+    BigDecimal aplicarDescuento(BigDecimal precioBase);
 }
