+++ Exp1_S3_Florencia_Morice/TiendaRopaOnline/src/main/java/com/tienda/discount/Decorator.java
@@ -1,14 +1,16 @@
 package com.tienda.discount;
 
+import java.math.BigDecimal;
+
 public abstract class Decorator implements Component {
     protected Component componente;
-
+    
     public Decorator(Component componente) {
         this.componente = componente;
     }
-
+    
     @Override
-    public double aplicarDescuento(double precioBase) {
+    public BigDecimal aplicarDescuento(BigDecimal precioBase) {
         return componente.aplicarDescuento(precioBase);
     }
 }
